Abstracted Dynamic Smile - Abstracted dynamic smile from submit happiness to individual component. - This allows for reuse on other pages if necessary.
