+++ happiness-frontend/src/pages/SubmitHappiness.js
@@ -3,6 +3,7 @@ import React, { useState } from "react";
 import "../App.css";
 import Journal from "../media/journal-icon.svg";
 import HappinessCommentModal from "../components/HappinessCommentModal";
+import DynamicSmile from "../components/DynamicSmile";
 
 export default function SubmitHappiness() {
   const [happiness, setHappiness] = useState(50);
@@ -57,16 +58,6 @@ export default function SubmitHappiness() {
   const submitHappiness = () => {
     //TODO submit happiness with backend.
   };
-  let happinessNumberEdges = (happiness) => {
-    //Saddest point: 200
-    //Happiest point: 160
-    return 190 - (happiness / 10) * 4;
-  };
-  let happinessNumberMiddle = (happiness) => {
-    //Saddest point: 160
-    //Happiest point: 200
-    return 150 + (happiness / 10) * 4;
-  };
 
   return (
     // Background:
@@ -94,27 +85,7 @@ export default function SubmitHappiness() {
         {/* Happy Face, Slider, and Happiness Number (Desktop only) */}
         <div className="flex flex-row items-center justify-center mobile-hidden">
           {/* Happy Face Decorator */}
-          <div className="flex-1 flex-row ">
-            <svg
-              width="90"
-              height="90"
-              viewBox="0 0 256 256"
-              className="mt-10 flex-1 mr-28"
-            >
-              <circle cx="128" cy="128" r="120" className="smile-head" />
-              <circle cx="98" cy="94" r="13" />
-              <circle cx="158" cy="94" r="13" />
-              <path
-                className="smile-mouth"
-                d={`M80,${happinessNumberEdges(
-                  happiness
-                )}, Q128,${happinessNumberMiddle(
-                  happiness
-                )} 176,${happinessNumberEdges(happiness)}`}
-              />
-            </svg>
-          </div>
-
+          <DynamicSmile happiness={happiness} />
           {/* Happiness Slider */}
           <input
             id="default-range"
